Description:
Added fortran 2003 examples; they compile if --enable-fortran2003 was used for the h5fc compiler. 
Enabled automatic check of the output.

Tested: jam (gmu ,intel, pgi)

+++ 1_8/FORTRAN/H5T/h5ex_t_vlstring.f90
@@ -20,8 +20,8 @@ PROGRAM main
   CHARACTER(LEN=18), PARAMETER :: filename  = "h5ex_t_vlstring.h5"
   CHARACTER(LEN=3) , PARAMETER :: dataset   = "DS1"
 
-  INTEGER(HSIZE_T) , PARAMETER :: dim0      = 4
-  INTEGER(SIZE_T)  , PARAMETER :: sdim      = 7
+  INTEGER(HSIZE_T), PARAMETER :: dim0      = 4
+  INTEGER(HSIZE_T), PARAMETER :: sdim      = 7
   INTEGER(HID_T)  :: file, filetype, space, dset ! Handles
   INTEGER :: hdferr
   INTEGER(HSIZE_T), DIMENSION(1:1) :: dims = (/dim0/)
